Tasks for building the workshop
