+++ .gitignore
@@ -1,5 +1,3 @@
 /build/
-/dist/
 /node_modules/
 /src/_book/
-/openlayers-workshop.zip
